change game pipline on callbacks
