+++ js/screen-greeting.js
@@ -1,4 +1,3 @@
-import controller from './controller';
 import renderTemplate from './render-template';
 import getFooter from './block-footer';
 
@@ -18,7 +17,7 @@ const contentTemplate = `
 </div>
 `;
 
-export default (state) => {
+export default (state, callback) => {
   const screen = document.createElement(`template`);
 
   const content = renderTemplate(contentTemplate);
@@ -29,7 +28,7 @@ export default (state) => {
 
   const link = screen.content.querySelector(`.greeting__continue`);
   link.addEventListener(`click`, () => {
-    controller(state, true);
+    callback(true);
   });
 
   return screen.content;
